TST/WIP: implement @slow decorator in numpy/lib/tests
Following the example in
http://doc.pytest.org/en/latest/example/simple.html#dynamically-adding-command-line-options

This adds a conftest.py in a test directories, numpy/lib/test, also
numpy/core/test. This is not scalable --- it will need to be
copy-pasted to all test folders.

One cannot add a conftest.py to the numpy root, because then pytest
tries to import numpy from its root folder, and it fails.

..